argparse: Remove top private declaration
Remove a declaration at the top stating that the internals of some of
the Formatter classes are an "implementation detail". This statement
here is unnecessary and incorrect. The declaration already exists with
each class. Formatter classes outside those listed also contain method
definitions that arguably would need to be included for completeness.